Change TopDown image orientation
Images are now oriented such that the robot's X axis points upwards from
the center of the image and the robot's Y axis points left from the
center of the image.

Previously, the robot's X axis pointed left from the center of the image
and the robot's Y axis pointed down from the center of the image.