Process images in grayscale instead of RGBA
RGBA was initially used to create pretty PNG images with transparent
pixels for areas outside of the view of the camera. This color space has
a large memory overhead compared to the original single-channel images.

The single-channel value of 0 is used to replace transparent pixels.
To roughly preserve the original image, pixels in the original image
with value 0 are set to 1.

Optimizations introduced in this commit roughly halve the runtime of
homography.proto_to_numpy.SpotImage.decoded_image_ground_plane() and
homography.perspective_transform.TopDown.get_view()