Remove same-geolocation restriction for anchor,pos
Allowing anchor,positive patches to be extracted from different
viewpoints allows different terrains in the same logical category (e.g.
concrete and speedway) to be associated with one another.

A consequence of this sampling strategy is that "viewpoint invariance"
as defined in VRL-PAP may not be as strongly learned. Instead, the model
learns broad categories of different terrains.